LayoutTests/editing/execCommand/break-out-of-empty-list-item-expected.txt

This tests CompositeEditCommand::breakOutOfEmptyListItem by inserting new paragraph

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS enterAtTarget('<ul><li>a <ul><li>b</li><li id="target"></li></ul> </li></ul>') is '<ul><li>a </li><ul><li>b</li></ul><li><br></li> </ul>'
PASS enterAtTarget('<ul><li>a <ul><li id="target"></li><li>b</li></ul> </li></ul>') is '<ul><li>a </li><li><br></li><ul><li>b</li></ul> </ul>'
PASS enterAtTarget('<ul><li>a <ul><li>b</li><li id="target"></li><li>c</li></ul> </li></ul>') is '<ul><li>a </li><ul><li>b</li></ul><li><br></li><ul><li>c</li></ul> </ul>'
PASS enterAtTarget('<ul><li>hello<ul><li id="target"><br></li></ul>world</li></ul>') is '<ul><li>hello<div><br></div>world</li></ul>'
PASS enterAtTarget('<ul><li>hello<ul><li id="target"><br></li></ul></li></ul>') is '<ul><li>hello</li><li><br></li></ul>'
PASS enterAtTarget('<ul><li><ul><li id="target"><br></li></ul>world</li></ul>') is '<ul><li><div><br></div>world</li></ul>'
PASS enterAtTarget('<ul><li><ul><li id="target"><br></li></ul></li></ul>') is '<ul><li></li><li><br></li></ul>'
PASS successfullyParsed is true

TEST COMPLETE
